Move 1.1.4 migrations to 1.1.6 and add an empty dir for 1.1.6 migrations to invoke the data migration
