+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v116/Migration.java
@@ -1,19 +1,16 @@
-package org.openmetadata.service.migration.postgres.v114;
+package org.openmetadata.service.migration.mysql.v116;
 
 import static org.openmetadata.service.migration.utils.V112.MigrationUtil.lowerCaseUserNameAndEmail;
 import static org.openmetadata.service.migration.utils.V114.MigrationUtil.fixTestSuites;
 
 import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
 import org.jdbi.v3.core.Handle;
 import org.openmetadata.service.jdbi3.CollectionDAO;
 import org.openmetadata.service.migration.api.MigrationProcessImpl;
 import org.openmetadata.service.migration.utils.MigrationFile;
 
-@Slf4j
 public class Migration extends MigrationProcessImpl {
   private CollectionDAO collectionDAO;
-  private Handle handle;
 
   public Migration(MigrationFile migrationFile) {
     super(migrationFile);
@@ -22,7 +19,6 @@ public class Migration extends MigrationProcessImpl {
   @Override
   public void initialize(Handle handle) {
     super.initialize(handle);
-    this.handle = handle;
     this.collectionDAO = handle.attach(CollectionDAO.class);
   }
